Add files via upload
A user needs to guess the number between 1 to 10 and enter a number. If user enters the correct guess number it displays "🎉 Congratulations! You guessed it right!". If the user guesses the number wrong it displays "The wheel spun... and landed on (guess number) and ❌ Sorry, better luck next time!"